+++ examples/benchmarks/gpu_copy_bw_performance.py
@@ -1,10 +1,10 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-"""Micro benchmark example for GPU SM copy bandwidth performance.
+"""Micro benchmark example for GPU copy bandwidth performance.
 
 Commands to run:
-  python3 examples/benchmarks/gpu_sm_copy_bw_performance.py
+  python3 examples/benchmarks/gpu_copy_bw_performance.py
 """
 
 from superbench.benchmarks import BenchmarkRegistry, Platform
@@ -12,11 +12,11 @@ from superbench.common.utils import logger
 
 if __name__ == '__main__':
     context = BenchmarkRegistry.create_benchmark_context(
-        'gpu-sm-copy-bw', platform=Platform.CUDA, parameters='--mem_type htoh htod dtoh dtod ptop'
+        'gpu-copy-bw', platform=Platform.CUDA, parameters='--mem_type htod dtoh dtod --copy_type sm dma'
     )
     # For ROCm environment, please specify the benchmark name and the platform as the following.
     # context = BenchmarkRegistry.create_benchmark_context(
-    #     'gpu-sm-copy-bw', platform=Platform.ROCM, parameters='--mem_type htoh htod dtoh dtod ptop'
+    #     'gpu-copy-bw', platform=Platform.ROCM, parameters='--mem_type htod dtoh dtod --copy_type sm dma'
     # )
 
     benchmark = BenchmarkRegistry.launch_benchmark(context)
