+++ superbench/benchmarks/micro_benchmarks/gpu_copy_performance/CMakeLists.txt
@@ -3,7 +3,7 @@
 
 cmake_minimum_required(VERSION 3.18)
 
-project(gpu_sm_copy LANGUAGES CXX)
+project(gpu_copy LANGUAGES CXX)
 
 find_package(CUDAToolkit QUIET)
 
@@ -12,9 +12,10 @@ if(CUDAToolkit_FOUND)
     message(STATUS "Found CUDA: " ${CUDAToolkit_VERSION})
 
     include(../cuda_common.cmake)
-    add_executable(gpu_sm_copy gpu_sm_copy.cu)
-    set_property(TARGET gpu_sm_copy PROPERTY CUDA_ARCHITECTURES ${NVCC_ARCHS_SUPPORTED})
-    install(TARGETS gpu_sm_copy RUNTIME DESTINATION bin)
+    set(CUDA_NVCC_FLAGS "${CUDA_NVCC_FLAGS} -lnuma")
+    add_executable(gpu_copy gpu_copy.cu)
+    set_property(TARGET gpu_copy PROPERTY CUDA_ARCHITECTURES ${NVCC_ARCHS_SUPPORTED})
+    install(TARGETS gpu_copy RUNTIME DESTINATION bin)
 else()
     # ROCm environment
     include(../rocm_common.cmake)
@@ -23,15 +24,16 @@ else()
         message(STATUS "Found ROCm: " ${HIP_VERSION})
 
         # Convert cuda code to hip code inplace
-        execute_process(COMMAND hipify-perl -inplace -print-stats gpu_sm_copy.cu
+        execute_process(COMMAND hipify-perl -inplace -print-stats gpu_copy.cu
                         WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/)
 
         # Add HIP targets
-        set_source_files_properties(gpu_sm_copy.cu PROPERTIES HIP_SOURCE_PROPERTY_FORMAT 1)
+        set_source_files_properties(gpu_copy.cu PROPERTIES HIP_SOURCE_PROPERTY_FORMAT 1)
         # Link with HIP
-        hip_add_executable(gpu_sm_copy gpu_sm_copy.cu)
+        set(HIP_HIPCC_FLAGS "${HIP_HIPCC_FLAGS} -lnuma")
+        hip_add_executable(gpu_copy gpu_copy.cu)
         # Install tergets
-        install(TARGETS gpu_sm_copy RUNTIME DESTINATION bin)
+        install(TARGETS gpu_copy RUNTIME DESTINATION bin)
     else()
         message(FATAL_ERROR "No CUDA or ROCm environment found.")
     endif()
