+++ superbench/common/devices/gpu.py
@@ -13,7 +13,6 @@ class GPU():
     def __init__(self):
         """Initilize."""
         self._vendor = self.get_vendor()
-        self._count = self.get_count()
         # TODO: check CUDA or ROCm availability accordingly
 
     def get_vendor(self):
@@ -36,22 +35,3 @@ class GPU():
     def vendor(self):
         """Get the GPU vendor."""
         return self._vendor
-
-    def get_count(self):
-        """Get GPU count.
-
-        Returns:
-            int: GPU count, nvidia or amd. 0 if no GPU found.
-        """
-        if 'nvidia' == self._vendor:
-            return len(list(Path('/dev').glob('nvidia[0-9]*')))
-        elif 'amd' == self._vendor:
-            bdf_prefix = '[0-9a-fA-F]' * 4 + ':' + '[0-9a-fA-F]' * 2 + ':' + '[0-9a-fA-F]' * 2 + '.' + '[0-9a-fA-F]*'
-            return len(list(Path('/sys/module/amdgpu/drivers/pci:amdgpu/').glob(bdf_prefix)))
-        else:
-            return 0
-
-    @property
-    def count(self):
-        """Get the GPU count."""
-        return self._count
