+++ superbench/config/default.yaml
@@ -63,20 +63,20 @@ superbench:
           proc_num: 8
           prefix: CUDA_VISIBLE_DEVICES={proc_rank} numactl -c $(({proc_rank}/2))
           parallel: yes
-    gpu-sm-copy-bw:
+    gpu-copy-bw:
       enable: true
       modes:
         - name: local
-          proc_num: 4
-          prefix: numactl -N {proc_rank} -m {proc_rank}
+          proc_num: 1
           parallel: no
       parameters:
         mem_type:
-          - htoh
           - htod
           - dtoh
           - dtod
-          - ptop
+        copy_type:
+          - sm
+          - dma
     kernel-launch:
       <<: *default_local_mode
     gemm-flops:
