--- tests/benchmarks/micro_benchmarks/test_gpu_sm_copy_bw_performance.py
@@ -1,77 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-"""Tests for disk-performance benchmark."""
-
-import numbers
-
-from tests.helper import decorator
-from superbench.benchmarks import BenchmarkRegistry, BenchmarkType, ReturnCode, Platform
-
-
-def _test_gpu_sm_copy_bw_performance_impl(platform):
-    """Test gpu-sm-copy-bw benchmark."""
-    benchmark_name = 'gpu-sm-copy-bw'
-    (benchmark_class,
-     predefine_params) = BenchmarkRegistry._BenchmarkRegistry__select_benchmark(benchmark_name, platform)
-    assert (benchmark_class)
-
-    size = 1048576
-    num_loops = 10000
-    mem_types = ['htoh', 'htod', 'dtoh', 'dtod', 'ptop']
-
-    parameters = '--mem_type %s --size %d --num_loops %d' % \
-        (' '.join(mem_types), size, num_loops)
-    benchmark = benchmark_class(benchmark_name, parameters=parameters)
-
-    # Check basic information
-    assert (benchmark)
-    ret = benchmark._preprocess()
-    assert (ret is True)
-    assert (benchmark.return_code == ReturnCode.SUCCESS)
-    assert (benchmark.name == benchmark_name)
-    assert (benchmark.type == BenchmarkType.MICRO)
-
-    # Check parameters specified in BenchmarkContext.
-    assert (benchmark._args.mem_type == mem_types)
-    assert (benchmark._args.size == size)
-    assert (benchmark._args.num_loops == num_loops)
-
-    # Check and revise command list
-    assert (len(benchmark._command_settings) == len(benchmark._commands))
-    for command_setting, command in zip(benchmark._command_settings, benchmark._commands):
-        devices = ' '.join(command_setting['devices'])
-        assert (command == '%s %s %d %d' % (benchmark._GpuSmCopyBwBenchmark__bin_path, devices, size, num_loops))
-        mem_type = command_setting['tag'].split('_')[0]
-        assert (mem_type in mem_types)
-
-    # Run benchmark
-    assert (benchmark._benchmark())
-
-    # Check results and metrics.
-    assert (benchmark.run_count == 1)
-    assert (benchmark.return_code == ReturnCode.SUCCESS)
-    assert (len(benchmark._command_settings) == len(benchmark.raw_data))
-    assert (len(benchmark._command_settings) == len(benchmark.result))
-    for idx, command_setting in enumerate(benchmark._command_settings):
-        raw_output_key = 'raw_output_%d' % idx
-        assert (raw_output_key in benchmark.raw_data)
-        assert (len(benchmark.raw_data[raw_output_key]) == 1)
-        assert (isinstance(benchmark.raw_data[raw_output_key][0], str))
-
-        output_key = command_setting['tag']
-        assert (output_key in benchmark.result)
-        assert (len(benchmark.result[output_key]) == 1)
-        assert (isinstance(benchmark.result[output_key][0], numbers.Number))
-
-
-@decorator.cuda_test
-def test_gpu_sm_copy_bw_performance_cuda():
-    """Test gpu-sm-copy-bw benchmark, CUDA case."""
-    _test_gpu_sm_copy_bw_performance_impl(Platform.CUDA)
-
-
-@decorator.rocm_test
-def test_gpu_sm_copy_bw_performance_rocm():
-    """Test gpu-sm-copy-bw benchmark, ROCm case."""
-    _test_gpu_sm_copy_bw_performance_impl(Platform.ROCM)
